WinUAE 4910

+++ src/od-fs/stubs.cpp
@@ -51,7 +51,7 @@ int get_guid_target (uae_u8 *out) {
     return 0;
 }
 
-uae_u8 *save_log (int bootlog, int *len) {
+uae_u8 *save_log (int bootlog, size_t *len) {
     STUB("");
     return NULL;
 }
@@ -259,7 +259,7 @@ void target_getdate(int *y, int *m, int *d)
     *d = 8;
 }
 
-uae_u8 *save_screenshot(int monid, int *len)
+uae_u8 *save_screenshot(int monid, size_t *len)
 {
     STUB("");
     return NULL;
